+++ src/Form/DayPicker/calendar.css
@@ -1,14 +1,14 @@
 /* DayPicker styles */
 
 .DayPicker {
-  display: inline-block;
-  /* background-color: red */
+  display: block;
+  /* background-color: red; */
 }
 
 .DayPicker-wrapper {
   position: relative;
   user-select: none;
-  padding-bottom: 1rem;
+  padding-bottom: 0;
   flex-direction: row;
 }
 
@@ -33,38 +33,36 @@
   position: absolute;
   cursor: pointer;
   top: 0;
-  color: #8b9898;
-  width: 1.25rem;
-  height: 1.25rem;
+  width: 20px;
+  height: 20px;
   display: inline-block;
-  background-size: 50%;
   background-repeat: no-repeat;
   background-position: center;
 }
 
-.DayPicker-NavButton:hover {
+.DayPicker-NavButton:hover:not(.DayPicker-NavButton--interactionDisabled) {
   opacity: 0.8;
 }
 
 .DayPicker-NavButton--prev {
   left: 12px;
-  background-image: url("./img/prev-arrow.svg");
+  background-image: url("./prev-arrow.svg");
 }
 
 .DayPicker-NavButton--next {
   right: 12px;
-  background-image: url("./img/next-arrow.svg");
+  background-image: url("./next-arrow.svg");
 }
 
 .DayPicker-NavButton--interactionDisabled {
-  opacity: 0.2;
+  opacity: 0;
 }
 
 .DayPicker-Caption {
   padding: 0 0.5rem;
   display: table-caption;
   text-align: center;
-  margin-bottom: 0.5rem;
+  margin-bottom: 24px;
 }
 
 .DayPicker-Caption > div {
@@ -76,7 +74,7 @@
 }
 
 .DayPicker-Weekdays {
-  margin-top: 1rem;
+  margin-top: 0;
   display: table-header-group;
 }
 
@@ -86,7 +84,6 @@
 
 .DayPicker-Weekday {
   display: table-cell;
-  padding: 0.5rem;
   font-weight: 900;
   line-height: 18px;
   font-size: 14px;
@@ -109,6 +106,7 @@
 }
 
 .DayPicker-Day {
+  position: relative;
   display: table-cell;
   padding: 0;
   vertical-align: top;
@@ -168,9 +166,15 @@
 }
 
 .DayPicker-Day--selected:not(.DayPicker-Day--disabled):not(.DayPicker-Day--outside) {
-  background: #effbff;
-  border: 0.8px solid #00ace2;
-  border-radius: 4px;
+  background: #f4f4f4;
+}
+.DayPicker-Day--selected:not(.DayPicker-Day--disabled):not(.DayPicker-Day--outside)
+  > .day:before {
+  content: url("./chosen.svg");
+  position: absolute;
+  z-index: 9;
+  top: 0;
+  left: 0;
 }
 
 .DayPicker-Day--selected:not(.DayPicker-Day--disabled):not(.DayPicker-Day--outside):hover {
@@ -178,9 +182,13 @@
 }
 
 .DayPicker:not(.DayPicker--interactionDisabled)
-  .DayPicker-Day:not(.DayPicker-Day--disabled):not(.DayPicker-Day--selected):not(.DayPicker-Day--outside):hover {
-  background-color: #f0f8ff;
-  border-radius: 0;
+  .DayPicker-Day:not(.DayPicker-Day--disabled):not(.DayPicker-Day--selected):not(.DayPicker-Day--outside):hover
+  > .day:before {
+  content: url("./chosen.svg");
+  position: absolute;
+  z-index: 9;
+  top: 0;
+  left: 0;
 }
 
 /* DayPickerInput */
@@ -199,4 +207,24 @@
   position: absolute;
   background: white;
   box-shadow: 0 2px 5px rgba(0, 0, 0, 0.15);
+}
+
+.day {
+  width: 48px;
+  min-height: 48px;
+}
+
+.day--date {
+  position: relative;
+  z-index: 10;
+  padding: 8px 0 0;
+}
+.day--price {
+  position: relative;
+  z-index: 100;
+  font-weight: 500;
+  line-height: 20px;
+  font-size: 10px;
+  text-align: center;
+  color: #00c455;
 }
